Rename projections script to modules
