+++ .github/workflows/deploy-2-start.yml
@@ -119,11 +119,11 @@ jobs:
         working-directory: cd
         run: python -m pip install -r scripts/spack_manifest/injection/requirements.txt
 
-      - name: Spack Manifest - Projection Injection
-        # Injects relevant projections into the spack manifest so users don't have to
+      - name: Spack Manifest - Modules Injection
+        # Injects relevant module projections and includes into the spack manifest so users don't have to
         # Note the cd prefix to the module name, as this is the path we cloned build-cd into
         run: |
-          python -m cd.scripts.spack_manifest.injection.projections \
+          python -m cd.scripts.spack_manifest.injection.modules \
             --manifest ${{ inputs.spack-manifest-path }} \
             --root-spec ${{ steps.manifest.outputs.root-spec-name }} \
             --packages "${{ vars.BUILD_DB_PACKAGES }}" \
