+++ scripts/spack_manifest/injection/modules.py
